+++ R/dbQueries-Results.R
@@ -5,21 +5,21 @@ NULL
 #' @import DBI
 #' @export
 setClass("DBIMockResult",
-  slots = c(
-    type = "character",
-    hash = "character",
-    path = "character",
-    statement = "character",
-    fetched = "logical"
-  ),
-  prototype = list(
-    type = NA_character_,
-    hash = NA_character_,
-    path = NA_character_,
-    statement = NA_character_,
-    fetched = FALSE
-  ),
-  contains = "DBIResult"
+         slots = c(
+           type = "character",
+           hash = "character",
+           path = "character",
+           statement = "character",
+           fetched = "logical"
+         ),
+         prototype = list(
+           type = NA_character_,
+           hash = NA_character_,
+           path = NA_character_,
+           statement = NA_character_,
+           fetched = FALSE
+         ),
+         contains = "DBIResult"
 )
 
 #' @rdname mock-db-methods
@@ -29,13 +29,24 @@ setMethod(
   "dbSendQuery", signature("DBIMockConnection", "character"),
   function(conn, statement, ...) {
     # create a new mock result with the type and a hash of the statement.
-    # TODO: extract the type from the statement instead of hard coding it.
     if (dittodb_debug_level(1)) {
       message(
         "Sending a query for the statement: \n", statement,
         "\nis being hased to: ", hash(statement)
       )
     }
+
+    # TDOO: if we are in expect_sql, then we should emit an error (or warning?) with SQL here
+    if (is_expecting()) {
+      stop(
+        "Fixture: ",
+        make_path(conn@path, get_type(statement), hash(statement)),
+        "\n",
+        clean_statement(statement),
+        call. = FALSE
+      )
+    }
+
     return(new(
       "DBIMockResult",
       type = get_type(statement),
@@ -88,22 +99,6 @@ setMethod(
   "dbHasCompleted",
   signature("DBIMockResult"),
   function(res, ...) return(invisible(TRUE))
-)
-
-#' @rdname mock-db-methods
-#' @importFrom methods setMethod new
-#' @export
-setMethod(
-  "dbGetQuery",
-  signature("DBIMockRPostgreSQLConnection", "character"),
-  function(conn, statement, ...) {
-    # TODO: this is really only needed for RPostgreSQL, and even for that, we
-    # likely could instead just mock `isPostgresqlIdCurrent` to return a valid
-    # value
-    # https://github.com/tomoakin/RPostgreSQL/blob/master/RPostgreSQL/R/PostgreSQLSupport.R#L266 #nolint
-    res <- dbSendQuery(conn, statement, ...)
-    return(mock_fetch(res, -1))
-  }
 )
 
 #' @rdname mock-db-methods