+++ NEWS.md
@@ -1,7 +1,7 @@
 # dittodb (development version)
 * ODBC connections that only specify a dsn now use the dsn as the path (@klmr, #132). 
-* Compatiblity for the forthcoming testthat 3e
-* Test changes for dbplyr
+* Compatibility for the forthcoming testthat 3e.
+* Test changes for {dbplyr} (@hadley, #134).
 
 # dittodb 0.1.1
 * Minor CRAN update that makes tests and examples conditional when `Suggests` packages are not available.